src/config deleted
